requirements-vercel.txt
# Vercel Serverless：减小体积、去掉无服务器环境无法使用的 Playwright 浏览器依赖
fastapi>=0.110,<1
python-dotenv>=1.0,<2
yt-dlp>=2024.1,<3000
httpx>=0.27,<1
curl-cffi>=0.7,<1
